autoramosweb/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
Django==4.0.5
django-celery-beat==2.3.0
django-crispy-forms==1.14.0
django-timezone-field==5.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imageio==2.19.5
kombu==5.2.4
lxml==4.9.0
numpy==1.23.1
outcome==1.2.0
Pillow==9.2.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
twill==3.0.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wsproto==1.1.0
